Refactor menu system into separate MenuManager module
